LayoutTests/platform/chromium-win/media/video-source-expected.txt

EVENT(loadstart)
EXPECTED (relativeURL(video.currentSrc) == 'content/test.ogv') OK
EXPECTED (source.getAttribute('src') == 'content/test.ogv') OK
EXPECTED (relativeURL(source.src) == 'content/test.ogv') OK
END OF TEST
